pyproject.toml
[project]
name = "nanobot-ai"
version = "0.1.3.post5"
description = "A lightweight personal AI assistant framework"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "nanobot contributors"}
]
keywords = ["ai", "agent", "chatbot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.9.0",
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "websockets>=12.0",
    "websocket-client>=1.6.0",
    "httpx[socks]>=0.25.0",
    "loguru>=0.7.0",
    "readability-lxml>=0.8.0",
    "rich>=13.0.0",
    "croniter>=2.0.0",
    # Legacy relay channels (disabled for enterprise)
    # "dingtalk-stream>=0.4.0",
    # "python-telegram-bot[socks]>=21.0",
    # "lark-oapi>=1.0.0",
    "socksio>=1.0.0",
    # Enterprise mobile app dependencies
    "qrcode>=7.4.0",
    "Pillow>=10.0.0",
    "PyJWT>=2.8.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
nanobot = "nanobot.cli.commands:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nanobot"]

[tool.hatch.build.targets.wheel.sources]
"nanobot" = "nanobot"

# Include non-Python files in skills
[tool.hatch.build]
include = [
    "nanobot/**/*.py",
    "nanobot/skills/**/*.md",
    "nanobot/skills/**/*.sh",
]

[tool.hatch.build.targets.sdist]
include = [
    "nanobot/",
    "README.md",
    "LICENSE",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
